+++ lib/whatsapp_sdk/resource/interactive_footer.rb
@@ -23,8 +23,7 @@ module WhatsappSdk
 
       sig { returns(T::Hash[T.untyped, T.untyped]) }
       def to_json
-        json = { text: text }
-        json
+        { text: text }
       end
 
       MAXIMUM_LENGTH = 60
@@ -41,9 +40,8 @@ module WhatsappSdk
         text_length = text.length
         return if text_length <= MAXIMUM_LENGTH
 
-        raise WhatsappSdk::Resource::Error::InvalidInteractiveFooter.new(
-          "invalid length #{text_length} for text in footer. Maximum length: #{MAXIMUM_LENGTH} characters."
-        )
+        raise WhatsappSdk::Resource::Error::InvalidInteractiveFooter,
+              "invalid length #{text_length} for text in footer. Maximum length: #{MAXIMUM_LENGTH} characters."
       end
     end
   end