+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # Unreleased 
+- Send interactive Reply Buttons Message @alienware https://github.com/ignacio-chiazzo/ruby_whatsapp_sdk/pull/79
 
 # v 0.7.3
 - Added the ability to reply messages. @alienware [#77](https://github.com/ignacio-chiazzo/ruby_whatsapp_sdk/pull/77)